Add testing of directories
